+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Add `NiimbotProtocol.getPaperInfo` to get paper dimensions from RFID (auto called and stored in `NiimbotAbstractClient.rfidInfo`).
 * Auto update RFID data on lid state change (for old printers).
 * Use `GetPrinterCapabilities` command to extend `NiimbotAbstractClient.printerInfo`.
+* Compact ambiguous version format.
+* Add `NiimbotAbstractClient.setSoundEnabled`, `NiimbotAbstractClient.printerInfo.settings`.
 
 # 0.46.0